Mock drs call when checking cloud platform
